Remove unused import metadata
