+++ core/__tests__/actions/records/records.ts
@@ -218,7 +218,7 @@ describe("actions/records", () => {
       );
 
       // Run the import task
-      await RecordOps.makeReadyAndCompleteImports();
+      await RecordOps.makeReady();
 
       connection.params = {
         csrfToken,
@@ -288,7 +288,7 @@ describe("actions/records", () => {
         }
       );
 
-      await RecordOps.makeReadyAndCompleteImports();
+      await RecordOps.makeReady();
 
       ({ records: invalidRecords, total: invalidTotal } =
         await specHelper.runAction<RecordsList>("records:list", connection));
