+++ core/__tests__/tasks/record/export.ts
@@ -11,7 +11,6 @@ import {
   Source,
   Property,
   GrouparooModel,
-  GroupMember,
 } from "../../../src";
 import { Op } from "sequelize";
 
@@ -236,8 +235,6 @@ describe("tasks/record:export", () => {
         await importB.reload();
         expect(importA.state).toBe("complete");
         expect(importB.state).toBe("complete");
-        expect(importA.processedAt).toBeTruthy();
-        expect(importB.processedAt).toBeTruthy();
       });
 
       test("it will append destinationIds from imports", async () => {
@@ -306,7 +303,9 @@ describe("tasks/record:export", () => {
       describe("with exportRecord", () => {
         let counter = 0;
 
-        beforeEach(() => {
+        beforeEach(async () => {
+          await Export.truncate();
+          await record.reload();
           counter = 0;
         });
 
@@ -321,7 +320,7 @@ describe("tasks/record:export", () => {
           });
         });
 
-        it("will not records not yet in the ready state", async () => {
+        it("will not export records not yet in the ready state", async () => {
           await record.update({ state: "pending" });
 
           await specHelper.runTask("record:export", {
@@ -329,46 +328,6 @@ describe("tasks/record:export", () => {
           }); // does not throw because it did not run
 
           expect(counter).toBe(0);
-        });
-
-        it("applies errors to the imports and export", async () => {
-          const run = await helper.factories.run();
-          const _import = await Import.create({
-            state: "importing",
-            creatorType: "run",
-            creatorId: run.id,
-            recordId: record.id,
-            data: {},
-            oldGroupIds: [],
-            newGroupIds: [group.id],
-          });
-
-          await _import.update({
-            state: "processing",
-            importedAt: new Date(),
-          });
-
-          await record.import();
-          await record.updateGroupMembership();
-
-          expect(_import.errorMessage).toBeFalsy();
-
-          // I don't throw, but append the error to the Export
-          await specHelper.runTask("record:export", {
-            recordId: record.id,
-          });
-
-          expect(counter).toBe(1);
-
-          await _import.reload();
-          expect(_import.state).toBe("failed");
-          expect(_import.errorMessage).toMatch(/oh no/);
-          const errorMetadata = JSON.parse(_import.errorMetadata);
-          expect(errorMetadata.message).toMatch(/oh no/);
-          expect(errorMetadata.step).toBe("record:export");
-          expect(errorMetadata.stack).toMatch(/RecordExport/);
-
-          await _import.destroy();
         });
       });
     });
