+++ core/src/models/GrouparooRecord.ts
@@ -29,7 +29,7 @@ import { Source } from "./Source";
 import { GrouparooModel } from "./GrouparooModel";
 import { ModelGuard } from "../modules/modelGuard";
 
-const STATES = ["draft", "pending", "ready", "deleted"] as const;
+const STATES = ["draft", "pending", "exporting", "ready", "deleted"] as const;
 
 const STATE_TRANSITIONS = [
   { from: "draft", to: "pending", checks: [] },
@@ -38,12 +38,30 @@ const STATE_TRANSITIONS = [
     from: "pending",
     to: "ready",
     checks: [
-      (instance: GrouparooRecord) =>
-        instance.validateRecordPropertiesAreReady(),
+      async (instance: GrouparooRecord) =>
+        await instance.validateRecordPropertiesAreReady(),
+    ],
+  },
+  {
+    from: "ready",
+    to: "exporting",
+    checks: [
+      async (instance: GrouparooRecord) =>
+        await instance.validateRecordPropertiesAreReady(),
     ],
   },
   { from: "ready", to: "pending", checks: [] },
+  {
+    from: "exporting",
+    to: "pending",
+    checks: [
+      async (instance: GrouparooRecord) =>
+        await instance.validateRecordPropertiesAreReady(),
+    ],
+  },
+  { from: "exporting", to: "draft", checks: [] },
   { from: "draft", to: "deleted", checks: [] },
+  { from: "exporting", to: "deleted", checks: [] },
   { from: "pending", to: "deleted", checks: [] },
   { from: "ready", to: "deleted", checks: [] },
 ];
@@ -145,8 +163,8 @@ export class GrouparooRecord extends LoggedModel<GrouparooRecord> {
     return RecordOps.removeProperties(this, properties);
   }
 
-  async buildNullProperties(state: GrouparooRecord["state"] = "pending") {
-    if (state !== "deleted") {
+  async buildNullProperties(state: RecordProperty["state"] = "pending") {
+    if (this.state !== "deleted") {
       return RecordOps.buildNullProperties([this], state);
     }
   }
@@ -185,7 +203,7 @@ export class GrouparooRecord extends LoggedModel<GrouparooRecord> {
 
   async export(
     force = false,
-    oldGroupsOverride?: Group[],
+    additionalGroups?: Group[],
     saveExports = true,
     sync = true,
     toDelete?: boolean
@@ -193,7 +211,7 @@ export class GrouparooRecord extends LoggedModel<GrouparooRecord> {
     return RecordOps._export(
       this,
       force,
-      oldGroupsOverride,
+      additionalGroups,
       saveExports,
       sync,
       toDelete
