+++ core/src/models/Import.ts
@@ -32,22 +32,13 @@ export interface ImportRecordProperties {
 
 const IMPORT_CREATORS = ["run"] as const;
 
-const STATES = [
-  "associating",
-  "importing",
-  "processing",
-  "failed",
-  "complete",
-] as const;
+const STATES = ["associating", "importing", "failed", "complete"] as const;
 
 const STATE_TRANSITIONS: StateMachine.StateTransition[] = [
   { from: "associating", to: "failed", checks: [] },
   { from: "associating", to: "importing", checks: [] },
   { from: "importing", to: "failed", checks: [] },
   { from: "importing", to: "complete", checks: [] },
-  { from: "importing", to: "processing", checks: [] },
-  { from: "processing", to: "failed", checks: [] },
-  { from: "processing", to: "complete", checks: [] },
 ];
 
 @Table({ tableName: "imports", paranoid: false })
@@ -81,17 +72,6 @@ export class Import extends CommonModel<Import> {
     this.setDataValue("data", JSON.stringify(value));
   }
 
-  @Column(DataType.TEXT)
-  // rawData: string;
-  get rawData(): ImportData {
-    //@ts-ignore
-    return JSON.parse(this.getDataValue("rawData") || "{}");
-  }
-  set rawData(value: ImportData) {
-    //@ts-ignore
-    this.setDataValue("rawData", JSON.stringify(value));
-  }
-
   @ForeignKey(() => GrouparooRecord)
   @Column
   recordId: string;
@@ -104,50 +84,6 @@ export class Import extends CommonModel<Import> {
   @Column
   createdRecord: boolean;
 
-  @Column(DataType.TEXT)
-  // oldRecordProperties: string;
-  get oldRecordProperties(): ImportRecordProperties {
-    //@ts-ignore
-    return JSON.parse(this.getDataValue("oldRecordProperties") || "{}");
-  }
-  set oldRecordProperties(value: ImportRecordProperties) {
-    //@ts-ignore
-    this.setDataValue("oldRecordProperties", JSON.stringify(value));
-  }
-
-  @Column(DataType.TEXT)
-  // newRecordProperties: string;
-  get newRecordProperties(): ImportRecordProperties {
-    //@ts-ignore
-    return JSON.parse(this.getDataValue("newRecordProperties") || "{}");
-  }
-  set newRecordProperties(value: ImportRecordProperties) {
-    //@ts-ignore
-    this.setDataValue("newRecordProperties", JSON.stringify(value));
-  }
-
-  @Column(DataType.TEXT)
-  // oldGroupIds: string;
-  get oldGroupIds(): string[] {
-    //@ts-ignore
-    return JSON.parse(this.getDataValue("oldGroupIds") || "[]");
-  }
-  set oldGroupIds(value: string[]) {
-    //@ts-ignore
-    this.setDataValue("oldGroupIds", JSON.stringify(value));
-  }
-
-  @Column(DataType.TEXT)
-  // newGroupIds: string;
-  get newGroupIds(): string[] {
-    //@ts-ignore
-    return JSON.parse(this.getDataValue("newGroupIds") || "[]");
-  }
-  set newGroupIds(value: string[]) {
-    //@ts-ignore
-    this.setDataValue("newGroupIds", JSON.stringify(value));
-  }
-
   @Column
   startedAt: Date;
 
@@ -156,9 +92,6 @@ export class Import extends CommonModel<Import> {
 
   @Column
   importedAt: Date;
-
-  @Column
-  processedAt: Date;
 
   @Column
   errorMessage: string;
@@ -171,10 +104,8 @@ export class Import extends CommonModel<Import> {
 
   async apiData() {
     const data = this.data || {};
-    const rawData = this.rawData || {};
 
     delete data._meta;
-    delete rawData._meta;
 
     const record = await this.$get("record");
 
@@ -190,21 +121,14 @@ export class Import extends CommonModel<Import> {
 
       //data
       data,
-      rawData,
 
       // lifecycle timestamps
       createdAt: APIData.formatDate(this.createdAt),
       startedAt: APIData.formatDate(this.startedAt),
       recordAssociatedAt: APIData.formatDate(this.recordAssociatedAt),
       importedAt: APIData.formatDate(this.importedAt),
-      processedAt: APIData.formatDate(this.processedAt),
 
-      // data before and after
       createdRecord: this.createdRecord,
-      oldRecordProperties: this.oldRecordProperties,
-      newRecordProperties: this.newRecordProperties,
-      oldGroupIds: this.oldGroupIds,
-      newGroupIds: this.newGroupIds,
 
       // errors
       errorMessage: this.errorMessage,
